+++ src/main/java/com/livinglogic/ul4/SetAST.java
@@ -67,6 +67,11 @@ public class SetAST extends CodeAST
 		super(template, posStart, posStop);
 	}
 
+	public List<SeqItemASTBase> getItems()
+	{
+		return items;
+	}
+
 	public void append(SeqItemASTBase item)
 	{
 		items.add(item);